Remove the Contributions section from README
The section claimed I will merge new implementations but the way the
library doc is written suggests I have no such intention. And indeed, I
do not. The `std` type are already in and implementing a supporting a
bunch of external crates would be a lot of work and even affect the
stability of this crate. It's better to do it the other way around.