Release version 0.1.5
There were a bunch of changes to the crate and the next step planned is
to release a 1.0 with semver trick. However the plan is also to remove
some things from 1.0 which would make the changelog confusing. Thus one
more release is going to be made before the final semver trick release.