Delete the helper functions & types
These are effectively scope creep and not required for the crate's main
goal. Having them around would burden library developers with auditing
them and compiling them. The crate should have as low barriers to use as
possible to encourage a wide range of libraries to use it and provide
the impls. Any additional functions were just hindering it.

That being said, the implementations will continue to live in 0.1.6
which is going to be semver-tricked. And long term, they are going to be
migrated to `configure_me` which, to my knowledge, is the only crate
that actually uses them. If some other crate wants them we can
collaborate and create a new shared crate later.