Remove impls for 128-bit atomics
These are now unstable however the feature gate `target_has_atomic` is
still turned on. This causes breakage on macOS.

Note that the removal doesn't break anything because the atomic was
never stable and it also doesn't break nightly because this crate didn't
opt into the nightly feature so it was broken even on nightly.